Move class in package belonging
