+++ Kitodo/src/main/java/org/kitodo/production/metadata/TreeNodeStruct3.java
@@ -15,8 +15,8 @@ import java.util.ArrayList;
 
 import org.kitodo.config.ConfigCore;
 import org.kitodo.config.enums.ParameterCore;
-import org.kitodo.helper.metadata.LegacyDocStructHelperInterface;
 import org.kitodo.production.helper.TreeNode;
+import org.kitodo.production.helper.metadata.legacytypeimplementations.LegacyDocStructHelperInterface;
 
 public class TreeNodeStruct3 extends TreeNode {
