+++ Kitodo/src/main/java/org/kitodo/production/metadata/copier/DestinationReferenceSelector.java
@@ -15,7 +15,7 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import org.apache.commons.configuration.ConfigurationException;
-import org.kitodo.helper.metadata.LegacyDocStructHelperInterface;
+import org.kitodo.production.helper.metadata.legacytypeimplementations.LegacyDocStructHelperInterface;
 
 /**
  * A DestinationReferenceSelector provides methods to retrieve document
