+++ Kitodo/src/main/java/org/kitodo/production/metadata/copier/MetadataSelector.java
@@ -12,7 +12,7 @@
 package org.kitodo.production.metadata.copier;
 
 import org.apache.commons.configuration.ConfigurationException;
-import org.kitodo.helper.metadata.LegacyDocStructHelperInterface;
+import org.kitodo.production.helper.metadata.legacytypeimplementations.LegacyDocStructHelperInterface;
 
 /**
  * Abstract base class that different types of metadata selectors are based on.
